Move scalars to common
