+++ server/Common/Exceptions/BookContainsErrorException.cs
@@ -1,6 +1,6 @@
 namespace Common.Exceptions;
 
-using Server.Domain.Scalars;
+using Common.Scalars.Types;
 
 public class BookContainsErrorException(string isbn, BookErrorType type)
     : Exception($"Book for {isbn} already contains error for type {type}") { }
